Linux build fix: Fix additional C++17 compatibility issues in tests and examples (#123)
Additional fixes for int to int64_t casting in:
- Test files for builders, type helpers, serialization
- Example files for stdio echo client/server
- All metadata and argument additions now use explicit int64_t cast

+++ examples/stdio_echo/stdio_echo_client_basic.cc
@@ -128,7 +128,7 @@ class BasicEchoClient : public echo::EchoClientBase {
       logInfo("Test 2: Request with params");
       auto params = make<Metadata>()
                         .add("message", "Hello, Echo Server!")
-                        .add("count", 42)
+                        .add("count", static_cast<int64_t>(42))
                         .build();
 
       auto future = sendRequest("echo.test", params);
@@ -163,7 +163,7 @@ class BasicEchoClient : public echo::EchoClientBase {
       for (int i = 0; i < 5; ++i) {
         auto params =
             make<Metadata>()
-                .add("index", i)
+                .add("index", static_cast<int64_t>(i))
                 .add(
                     "timestamp",
                     std::chrono::system_clock::now().time_since_epoch().count())
@@ -222,7 +222,7 @@ class BasicEchoClient : public echo::EchoClientBase {
     for (int i = 0; i < 3; ++i) {
       std::this_thread::sleep_for(std::chrono::seconds(1));
 
-      auto ping_params = make<Metadata>().add("count", i + 1).build();
+      auto ping_params = make<Metadata>().add("count", static_cast<int64_t>(i + 1)).build();
 
       sendNotification("ping", ping_params);
     }
